# Remove IsFastSort optimizations from String