--- a/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.Unix.cs
+++ b/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.Unix.cs
@@ -365,6 +365,10 @@ namespace System.Globalization
                             *matchLengthPtr = target.Length;
                         return i;
                     }
+
+                    // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                    if (*(a + sourceIndex) >= 0x80)
+                        goto InteropCall;
                 }
 
                 return -1;
@@ -439,6 +443,10 @@ namespace System.Globalization
                             *matchLengthPtr = target.Length;
                         return i;
                     }
+
+                    // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                    if (*(a + sourceIndex) >= 0x80)
+                        goto InteropCall;
                 }
 
                 return -1;
@@ -561,7 +569,7 @@ namespace System.Globalization
                     length--;
                 }
 
-                // The match can be invalidated by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
 
                 if (source.Length < prefix.Length)
                 {
@@ -614,7 +622,7 @@ namespace System.Globalization
                     length--;
                 }
 
-                // The match can be invalidated by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
 
                 if (source.Length < prefix.Length)
                 {
@@ -703,7 +711,7 @@ namespace System.Globalization
                     length--;
                 }
 
-                // The match can be invalidated by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
 
                 if (source.Length < suffix.Length)
                 {
@@ -756,7 +764,7 @@ namespace System.Globalization
                     length--;
                 }
 
-                // The match can be invalidated by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
+                // The match can be affected by following NonSpacingMark, etc. Verify that the character following the match is regular ASCII.
 
                 if (source.Length < suffix.Length)
                 {