--- a/src/classlibnative/bcltype/stringnative.h
+++ b/src/classlibnative/bcltype/stringnative.h
@@ -44,9 +44,6 @@ public:
     //
     // Search/Query Methods
     //
-    static FCDECL1(FC_BOOL_RET, IsFastSort, StringObject* pThisRef);
-    static FCDECL1(FC_BOOL_RET, IsAscii, StringObject* pThisRef);
-
     static FCDECL6(INT32, CompareOrdinalEx, StringObject* strA, INT32 indexA, INT32 countA, StringObject* strB, INT32 indexB, INT32 countB);
 
     static FCDECL2(FC_CHAR_RET, GetCharAt, StringObject* pThisRef, INT32 index);