… src/utilcode/util_nodependencies.cpp
… src/vm/ecalllist.h
… src/vm/object.cpp