@jkotas 2019-09-18
Performance results:

The current microbenchmarks in performance repo are running the same tests on the same string, so they are helped by the cached state. For example, results for `System.Tests.Perf_String.IndexOf`:

Before:

|  Method |              options |       Mean |     Error |    StdDev |     Median |        Min |        Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|------:|------:|------:|----------:|
| IndexOf |    OrdinalIgnoreCase | 131.131 us | 0.8094 us | 0.7175 us | 130.867 us | 130.497 us | 132.775 us |     - |     - |     - |         - |

After:

|  Method |              options |       Mean |     Error |    StdDev |     Median |        Min |        Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|------:|------:|------:|----------:|
| IndexOf |    OrdinalIgnoreCase | 134.921 us | 0.7286 us | 0.6459 us | 135.016 us | 134.007 us | 135.810 us |     - |     - |     - |         - |

The scenario tested by the microbenchmarks is not realistic. It is unlikely that real workloads are going to be running the same IndeOf on the two exact same string instances repeatedly. When the microbenchmark is modified to create a fresh instance of string using `string.Copy`, the version in this PR is faster.

Before:

|  Method |              options |       Mean |     Error |    StdDev |     Median |        Min |        Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|-------:|------:|------:|----------:|
| IndexOf |    OrdinalIgnoreCase | 139.426 us | 1.4603 us | 1.3660 us | 138.840 us | 137.604 us | 142.475 us | 1.6741 |     - |     - |  19.83 KB |

After:

|  Method |              options |       Mean |     Error |    StdDev |     Median |        Min |        Max |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|-------:|-------:|------:|----------:|
| IndexOf |    OrdinalIgnoreCase | 138.191 us | 0.9291 us | 0.8237 us | 138.240 us | 136.397 us | 139.309 us | 1.6304 |      - |     - |  19.83 KB |

@adamsitnik Is it worth it to update the micro-benchmarks for IndexOf and similar operations to create string clones to be more realistic?