@adamsitnik 2019-09-18
>  Is it worth it to update the micro-benchmarks for IndexOf and similar operations to create string clones to be more realistic?

Yes. I am going to implement more realistic benchmarks https://github.com/dotnet/performance/issues/885