@adamsitnik 2019-09-18
The Windows failures seems to be not related. I am going to re-run the tests

/azp run