@adamsitnik 2019-09-18
The Linux failures might have exposed differences between the `string` and `ReadOnlySpan<char>` text operations. I am going to try one thing to see if this is true.