@jkotas 2019-09-19
@tarekgh The following test case is failing with my changes:

```
IsPrefix(compareInfo: CompareInfo - , source: "o\u0308", value: "o", options: None, expected: False)
```

We return true with my change because of `o` matches and it is in the ASCII set. We never look at `\u0308`. How is this supposed to work? Do we need to look at the following character in source, even after we find a match?

The same bug exists in Span version of StartsWith. We have test coverage hole for the Span version of StartsWith.