@tarekgh 2019-09-19
@jkotas  

> How is this supposed to work? Do we need to look at the following character in source, even after we find a match?

```ö``` is not equal to ```o``` so the expected result is right (which is false). We'll need to look at the next character. we can look if the next character is NonSpacingMark character but still this may not be enough as we can encounter some characters with no sorting weight which I'll not be sure what would be the behavior (I can try that). Instead, I suggest if the next char is not Ascii, then just fallback to the slow path. If you want, I can try to think in a better way if we need to avoid going to the slow path.