@jkotas 2019-09-19
>  I suggest if the next char is not Ascii, then just fallback to the slow path.

I think that's fine. Thanks!