@jkotas 2019-09-20
I am fixing the test coverage hole for Span-based globalization methods in https://github.com/dotnet/corefx/pull/41224