@adamsitnik 2019-09-23
@jkotas could you please wait with merging this PR until I finish https://github.com/dotnet/performance/pull/892 ? I just want to make sure that this change is tracked properly in our performance reporting system. It should be a matter of hours from now