@adamsitnik 2019-09-23
@jkotas thanks for reviewing https://github.com/dotnet/performance/pull/892 , I've merged it just right now. We run all benchmarks every 4 hours and it takes +-1h to run all Unix benchmarks (they are distributed to multiple machines). So 5h from now we can merge this PR. I hope it's not a problem!